+++ editors/helix/queries/postgresql-conf/textobjects.scm
@@ -1,0 +1,15 @@
+(setting) @entry.around
+
+(setting
+  value: (_) @entry.inside)
+
+(hba_rule) @entry.around
+
+(user_mapping) @entry.around
+
+(auth_option) @parameter.around
+
+(auth_option
+  value: (_) @parameter.inside)
+
+(comment)+ @comment.around