+++ editors/zed/languages/postgresql-conf/outline.scm
@@ -1,0 +1,11 @@
+(setting
+  name: (setting_name) @name) @item
+
+(hba_rule
+  type: (connection_type) @context
+  database: (database) @name
+  user: (user) @context.extra) @item
+
+(user_mapping
+  map: (map_name) @name
+  system_user: (_) @context) @item