+++ editors/zed/languages/postgresql-conf/textobjects.scm
@@ -1,0 +1,7 @@
+(setting) @function.around
+
+(hba_rule) @function.around
+
+(user_mapping) @function.around
+
+(comment)+ @comment.around